+++ back/src/main/java/com/openclassrooms/mddapi/repository/UserRepository.java
@@ -1,17 +1,16 @@
 package com.openclassrooms.mddapi.repository;
 
-import com.openclassrooms.mddapi.models.UserEntity;
+import com.openclassrooms.mddapi.models.entities.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
   Optional<UserEntity> findByLoginOrEmail(final String login, final String email);
 
   Boolean existsByLoginOrEmail(final String login, final String email);
-
 
 }
