+++ back/src/main/java/com/openclassrooms/mddapi/services/UserService.java
@@ -1,6 +1,6 @@
 package com.openclassrooms.mddapi.services;
 
-import com.openclassrooms.mddapi.models.UserEntity;
+import com.openclassrooms.mddapi.models.entities.UserEntity;
 import com.openclassrooms.mddapi.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
@@ -12,11 +12,11 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public void delete(Long id) {
+    public void delete(Integer id) {
         this.userRepository.deleteById(id);
     }
 
-    public UserEntity findById(Long id) {
+    public UserEntity findById(Integer id) {
         return this.userRepository.findById(id).orElse(null);
     }
 }
